% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_typecode.r
\name{flash_typecode}
\alias{flash_typecode}
\title{flash_typecode}
\usage{
flash_typecode(file_type_code, odbc)
}
\arguments{
\item{file_type_code}{file_type_code of file from input_file_type table}

\item{odbc}{ODBC connector containing destination database details}
}
\description{
Check for any waiting files per typecode specified in
the input_file_type table and add them to the input_file
table
}
