% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flick_adhoc_csv.r
\name{flick_adhoc_csv}
\alias{flick_adhoc_csv}
\title{flick_adhoc_csv}
\usage{
flick_adhoc_csv(file_path, odbc, schema, overwrite, append)
}
\arguments{
\item{file_path}{Path to the file you want to load}

\item{odbc}{ODBC connector containing destination database details}

\item{schema}{Schema in the destination database that you want to load the data to}

\item{overwrite}{Overwrite existing table}

\item{append}{Append data to table}
}
\description{
Load any single CSV file to a SQL table in the specified schema of the
database / server specified in the ODBC connector.
}
\details{
The table name assigned is the file name.
An ODBC connector must be setup, and the schema passed to the schema
parameter must exist in the destination database
}
