% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aux_tables.r
\name{init_aux_tables}
\alias{init_aux_tables}
\title{init_aux_tables}
\usage{
init_aux_tables(odbc)
}
\arguments{
\item{odbc}{ODBC connector containing destination database details}
}
\description{
Set up the auxiliary metadata tables in the specified
database / server specified in the ODBC connector.
}
\details{
An ODBC connector must be setup, and the schema passed to the schema
parameter must exist in the destination database
}
