% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flick_csv_folder.R
\name{flick_csv_folder}
\alias{flick_csv_folder}
\title{flick_csv_folder}
\usage{
flick_csv_folder(folder, odbc, tablename, schema)
}
\arguments{
\item{folder}{Path to the folder of files you want to load}

\item{odbc}{ODBC connector containing destination database details}

\item{tablename}{Table in the destination database that you want to load the data to}

\item{schema}{Schema in the destination database that you want to load the data to}
}
\description{
Load any folder of  CSV files to a SQL table in the specified schema of the
database / server specified in the ODBC connector. All files must be the same structure
}
\details{
The table name assigned is the file name. of the first file loaded. The file names are also
written to the table
An ODBC connector must be setup, and the schema that is  passed to the schema
parameter must exist in the destination database
}
